+++ src/room007/models/model.py
@@ -1,6 +1,8 @@
 # vim: set fileencoding=utf-8 :
 import itertools
+import abc
 
+from room007.data import feature_data
 
 class Option(object):
     def __init__(self, options, default):
@@ -45,41 +47,17 @@ class OptionsSetter(object):
             return [('||| no parameter combinations |||', {option_name: option.default for option_name, option in self.options.items()})]
 
 
-class FeatureManager(object):
-    save = False
-    def __init__(self, fun=None):
-        self.fun = fun
-
-    def __call__(self, *args, **kwargs):
-        data = args[0]
-        feature_name = self._get_feature_name(args, kwargs)
-        if feature_name in data:
-            features = data[feature_name].apply(lambda x: eval(x) if '[' in x else x)
-        else:
-            features = self.fun(*args, **kwargs)
-        if self.save:
-            data[feature_name] = features
-        return features
-
-    def _get_feature_name(self, args, kwargs):
-        """ get the feature name for a function, dependent on the argument and kwargs
-        args & kwargs are not expanded because they are treated as single arguments
-        """
-        feature_name = self.fun.__name__
-        if len(args)>1:
-            feature_name += '_'
-            feature_name += '_'.join(args[1:])
-        if kwargs:
-            feature_name += '_'
-            feature_name += '_'.join([key+':'+value for key, value in sorted(kwargs.items())])
-        return feature_name
-
 class Features(object):
-    def save_features_to_dataframe(self, data):
-        FeatureManager.save = True
-        self._get_data_independent_features(data)
-        FeatureManager.save = False
+    __metaclass__ = abc.ABCMeta
 
+    def save_features_to_dataframe(self, data):
+        feature_data.FeatureManager.save = True
+        self._get_data_independent_features(data)
+        feature_data.FeatureManager.save = False
+
+    @abc.abstractmethod
+    def _get_data_independent_features(self, data):
+        pass
 
 class Predictor(object):
     OptionsSetter = OptionsSetter
